refactor: extract layout memoization logic into useChartLayout hook
Extracts the large block of layout-related `useMemo` hooks from `ChartContainer.tsx` into a new custom hook `useChartLayout`. This significantly reduces the size and complexity of the main component file, improving readability and maintainability without altering runtime behavior. Also cleans up unused imports in the main file.
